requirements.txt
fastmcp>=0.9.0
azure-identity>=1.15.0
httpx>=0.27.0
openai>=1.30.0
python-dotenv>=1.0.0
jinja2>=3.1.0
pandas>=2.0.0
numpy>=2.0.0